heatmap added
Added a quick heatmap to suss out correlation.  Right now it's picking out some obvious correlations though, for example Biden votes correlate with COVID cases but this is because everything is in absolute numbers, we need to get cases per 100k in the data because we expect higher population density areas to both have more cases overall and to swing Democratic.